Add authorization checks #12

+++ index.js
@@ -7,32 +7,21 @@ const db = new PGClient()
 db.connectSync(process.env.CONNECTION_STRING)
 
 
-
 const client = mqtt.connect(process.env.BROKER_URL)
 
 MqttRequest.timeout = 5000;
-
-db.connectSync(process.env.CONNECTION_STRING)
-
 
 /** @type {MqttRequest}*/
 export const mqttReq = new MqttRequest.default(client);
 
 console.log(`Broker URL: ${process.env.BROKER_URL}`)
 
-mqttReq.response("demo", payload => {
-    payload = JSON.parse(payload)
-    payload.message += " and hi from scheduling-service"
-    console.log(payload)
-    return JSON.stringify(payload)
-})
-
 mqttReq.response("v1/dentists/read", (payload) => {
     payload = JSON.parse(payload)
 
     try {
         const dentists = db.querySync("SELECT id, name, clinic_id FROM public.user WHERE role = 'dentist'")
-        return JSON.stringify({ httpStatus: 200, dentists})
+        return JSON.stringify({ httpStatus: 200, dentists })
     } catch (e) {
         return JSON.stringify({ httpStatus: 500, message: `Some error occurred` })
     }
@@ -40,9 +29,17 @@ mqttReq.response("v1/dentists/read", (payload) => {
 
 mqttReq.response("v1/timeslots/delete", (payload) => {
     payload = JSON.parse(payload)
+    const token = jwt.decode(payload.token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
+    if (token.role !== 'dentist') {
+        return JSON.stringify({ httpStatus: 403, message: 'Forbidden' });
+    }
 
     try {
-        const token = jwt.decode(payload.token)
         const result = db.querySync(
             'DELETE FROM public.timeslot WHERE id = $1 AND dentist_id = $2 RETURNING *', [payload.timeslotId, token.id]
         );
@@ -60,8 +57,18 @@ mqttReq.response("v1/timeslots/delete", (payload) => {
 mqttReq.response("v1/timeslots/create", (payload) => {
     payload = JSON.parse(payload);
 
+    const token = jwt.decode(payload.token)
+    console.log(token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
+    if (token.role !== 'dentist') {
+        return JSON.stringify({ httpStatus: 403, message: 'Forbidden' });
+    }
+
     try {
-        const token = jwt.decode(payload.token)
         db.querySync("INSERT INTO public.timeslot (dentist_id, start_time, end_time) VALUES ($1, $2, $3)", [token.id, payload.start_time, payload.end_time]);
 
         const insertedTimeslot = db.querySync("SELECT * FROM public.timeslot WHERE dentist_id = $1 AND start_time = $2 AND end_time = $3", [token.id, payload.start_time, payload.end_time]);
@@ -70,7 +77,7 @@ mqttReq.response("v1/timeslots/create", (payload) => {
             return JSON.stringify({
                 httpStatus: 201,
                 message: `Created a new timeslot from ${payload.start_time} to ${payload.end_time}`,
-                timeslot: insertedTimeslot[0] 
+                timeslot: insertedTimeslot[0]
             });
         }
     } catch (e) {
@@ -84,18 +91,25 @@ mqttReq.response("v1/timeslots/create", (payload) => {
 
 mqttReq.response("v1/dentists/ratings/create", (payload) => {
     payload = JSON.parse(payload)
+
+    const token = jwt.decode(payload.token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
     try {
         console.log(typeof payload.rating)
         console.log(typeof payload.favorite_dentist)
-        if (typeof payload.rating != "number" && payload.rating != undefined){
-            return JSON.stringify({ httpStatus: 400, message: "Rating has to be a number"})
+        if (typeof payload.rating != "number" && payload.rating != undefined) {
+            return JSON.stringify({ httpStatus: 400, message: "Rating has to be a number" })
         }
-        
-        if (typeof payload.favorite_dentist != "boolean" && payload.favorite_dentist != undefined){
-            return JSON.stringify({ httpStatus: 400, message: "Favorite_dentist field has to have a boolean value"})
+
+        if (typeof payload.favorite_dentist != "boolean" && payload.favorite_dentist != undefined) {
+            return JSON.stringify({ httpStatus: 400, message: "Favorite_dentist field has to have a boolean value" })
         }
-        if (payload.rating < 1 || payload.rating > 5){
-            return JSON.stringify({ httpStatus: 400, message: "Rating has to be a number between 1 and 5"})
+        if (payload.rating < 1 || payload.rating > 5) {
+            return JSON.stringify({ httpStatus: 400, message: "Rating has to be a number between 1 and 5" })
         }
         //if there is not dentist with the payload.dentistId -> insert new dentist and new rating
         //don't forget to also send the patient_id -> this table has a compound key of patient_id and dentist_id
@@ -105,31 +119,31 @@ mqttReq.response("v1/dentists/ratings/create", (payload) => {
             return JSON.stringify({ httpStatus: 400, message: "You can't post a rating for a patient or add a patient to your favorites" });
         }
         const dentist_on_user = db.querySync(`SELECT * FROM public.patient_on_dentist WHERE dentist_id = $1 and patient_id = $2`, [payload.dentistId, token.id]);
-        if (dentist_on_user.length === 0){
-            if (payload.rating != undefined && payload.favorite_dentist != undefined){
+        if (dentist_on_user.length === 0) {
+            if (payload.rating != undefined && payload.favorite_dentist != undefined) {
                 const rating = parseInt(payload.rating)
                 db.querySync("insert into public.patient_on_dentist (patient_id, dentist_id, rating, favorite_dentist) values ($1, $2, $3, $4)", [token.id, payload.dentistId, rating, payload.favorite_dentist])
-            } else if (payload.rating != undefined){
+            } else if (payload.rating != undefined) {
                 const rating = parseInt(payload.rating)
                 db.querySync("insert into public.patient_on_dentist (patient_id, dentist_id, rating) values ($1, $2, $3)", [token.id, payload.dentistId, rating])
-            } else if (payload.favorite_dentist != undefined){
+            } else if (payload.favorite_dentist != undefined) {
                 db.querySync("insert into public.patient_on_dentist (patient_id, dentist_id, favorite_dentist) values ($1, $2, $3)", [token.id, payload.dentistId, payload.favorite_dentist])
             }
         }
         //if there is a dentist in the dentist_on_patient table, just insert a new rating.
-        else{
-            if (payload.rating != undefined && payload.favorite_dentist != undefined){
+        else {
+            if (payload.rating != undefined && payload.favorite_dentist != undefined) {
                 db.querySync(`UPDATE public.patient_on_dentist SET rating = $1, favorite_dentist = $2 where dentist_id = $3`, [payload.rating, payload.favorite_dentist, payload.dentistId])
             }
-            else if (payload.rating != undefined){
+            else if (payload.rating != undefined) {
                 db.querySync(`UPDATE public.patient_on_dentist SET rating = $1 where dentist_id = $2`, [payload.rating, payload.dentistId])
-            } else if (payload.favorite_dentist != undefined){
+            } else if (payload.favorite_dentist != undefined) {
                 db.querySync(`UPDATE public.patient_on_dentist SET favorite_dentist = $1 where dentist_id = $2`, [payload.favorite_dentist, payload.dentistId])
             }
-             
+
         }
         const dentist = db.querySync(`select * from public.patient_on_dentist where patient_id = $1 and dentist_id = $2`, [token.id, payload.dentistId])
-        return JSON.stringify({ httpStatus: 201, dentist})
+        return JSON.stringify({ httpStatus: 201, dentist })
     } catch (e) {
         console.log(e)
         return JSON.stringify({ httpStatus: 500, message: `Some error occurred`, errorInternal: e })
@@ -139,17 +153,16 @@ mqttReq.response("v1/dentists/ratings/create", (payload) => {
 
 mqttReq.response("v1/clinics/read", (payload) => {
     payload = JSON.parse(payload)
-    console.log(payload)
 
     try {
         const clinics = db.querySync('SELECT * FROM public.clinic')
-        return JSON.stringify({ httpStatus: 200, clinics})
+        return JSON.stringify({ httpStatus: 200, clinics })
     } catch (e) {
-        return JSON.stringify({ httpStatus: 400, message: "No clinics found"})
+        return JSON.stringify({ httpStatus: 400, message: "No clinics found" })
     }
 })
 mqttReq.response("v1/timeslots", (payload) => {
-    
+
     payload = JSON.parse(payload);
     console.log(payload)
 
@@ -160,30 +173,25 @@ mqttReq.response("v1/timeslots", (payload) => {
         db.querySync("select from public.timeslot where start_time = $1", [payload.startTime || Date.now()])
         return JSON.stringify({ httpStatus: 201, message: `${payload}` })
     } catch (e) {
-        return JSON.stringify({ httpStatus: 400, message: "Timeslots with this start time not found."})
-    }
-});
-
-mqttReq.response("v1/appointments/all", (payload) => {
-    payload = JSON.parse(payload)
-
-    try {
-        const token = jwt.decode(payload.token)
-        const appointments = db.querySync(`SELECT appointment.id AS appointment_id, appointment.*, timeslot.*, "user".name AS patient_name
-        FROM appointment
-        JOIN timeslot ON timeslot.id = appointment.timeslot_id
-        JOIN "user" ON "user".id = appointment.patient_id
-        WHERE appointment.dentist_id = $1`, [token.id])
-        return JSON.stringify({ httpStatus: 200, appointments})
-    } catch (e) {
-        return JSON.stringify({ httpStatus: 500, message: `Some error occurred` })
+        return JSON.stringify({ httpStatus: 400, message: "Timeslots with this start time not found." })
     }
 });
 
 mqttReq.response("v1/appointments/create", (payload) => {
     payload = JSON.parse(payload);
-    console.log('Received payload:', payload);
-        if (!payload.patient_id|| !payload.dentist_id || !payload.timeslot_id)
+
+    const token = jwt.decode(payload.token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
+    if (token.role !== 'user') {
+        return JSON.stringify({ httpStatus: 403, message: 'Forbidden' });
+    }
+
+
+    if (!payload.patient_id || !payload.dentist_id || !payload.timeslot_id)
         return JSON.stringify({ httpStatus: 400, message: "Patient ID, Dentist ID, Timeslot ID do not exist" })
 
     try {
@@ -191,7 +199,7 @@ mqttReq.response("v1/appointments/create", (payload) => {
             "INSERT INTO public.appointment (patient_id, dentist_id, timeslot_id, cancelled, confirmed) VALUES ($1, $2, $3, $4, $5) RETURNING *",
             [payload.patient_id, payload.dentist_id, payload.timeslot_id, payload.cancelled, payload.confirmed]
         );
-        const appointment= result[0]
+        const appointment = result[0]
         return JSON.stringify({ httpStatus: 201, appointment, message: `Appointment created` });
     } catch (error) {
         console.error('Error processing appointment creation:', error);
@@ -201,7 +209,15 @@ mqttReq.response("v1/appointments/create", (payload) => {
 
 mqttReq.response("v1/appointments/read", (payload) => {
     payload = JSON.parse(payload);
-    console.log('Received payload:', payload);
+
+    const token = jwt.decode(payload.token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
+    // TODO: only let the user see their own appointments
+
     const appointmentId = parseInt(payload.appointmentId, 10);
     if (isNaN(appointmentId)) {
         return JSON.stringify({ httpStatus: 400, message: 'Invalid payload. Appointment ID is not a valid number.' });
@@ -221,60 +237,33 @@ mqttReq.response("v1/appointments/read", (payload) => {
 
 mqttReq.response("v1/appointments/update", (payload) => {
     payload = JSON.parse(payload);
-    console.log('Received payload:', payload);
+
+    const token = jwt.decode(payload.token)
+
+    if (!token) {
+        return JSON.stringify({ httpStatus: 401, message: 'Unauthorized' });
+    }
+
+    // TODO: only let the user update their own appointments
+
     const appointmentId = parseInt(payload.appointmentId, 10);
     const requestBody = payload.requestBody;
-
-    //check if appointment id is an integer
     if (isNaN(appointmentId)) {
         return JSON.stringify({ httpStatus: 400, message: 'Invalid payload. Appointment ID is not a valid number.' });
     }
-
     try {
         const currentAppointment = db.querySync('SELECT * FROM public.appointment WHERE id = $1', [appointmentId]);
-
         if (currentAppointment.length === 0) {
             return JSON.stringify({ httpStatus: 404, message: `Appointment with ID ${appointmentId} not found.` });
         }
-
-        const updateFields = [];
-        const updateValues = [];
-
-        // Check and add fields to be updated
-        if (requestBody.patient_id !== undefined) {
-            updateFields.push(`patient_id = ${requestBody.patient_id}`);
-        }
-
-        if (requestBody.dentist_id !== undefined) {
-            updateFields.push(`dentist_id = ${requestBody.dentist_id}`);
-        }
-
-        if (requestBody.timeslot_id !== undefined) {
-            updateFields.push(`timeslot_id = ${requestBody.timeslot_id}`);
-        }
-
-        if (requestBody.cancelled !== undefined) {
-            updateFields.push(`cancelled = ${requestBody.cancelled}`);
-        }
-
-        if (requestBody.confirmed !== undefined) {
-            updateFields.push(`confirmed = ${requestBody.confirmed}`);
-        }
-
-        if (updateFields.length === 0) {
-            return JSON.stringify({ httpStatus: 400, message: 'No fields provided for update.' });
-        }
-
-        //update fields in the database
-        const updateQuery = `UPDATE public.appointment SET ${updateFields.join(', ')} WHERE id = $1 RETURNING *`;
-
-        const result = db.querySync(updateQuery, [appointmentId]);
+        const result = db.querySync(
+            'UPDATE public.appointment SET patient_id = $1, dentist_id = $2, timeslot_id = $3, cancelled = $4, confirmed = $5 WHERE id = $6 RETURNING *',
+            [requestBody.patient_id, requestBody.dentist_id, requestBody.timeslot_id, requestBody.cancelled, requestBody.confirmed, appointmentId]
+        );
         const updatedAppointment = result.length > 0 ? result[0] : null;
-
         if (!updatedAppointment) {
             return JSON.stringify({ httpStatus: 500, message: 'Failed to retrieve updated appointment.' });
         }
-
         console.log('Updated appointment:', updatedAppointment);
         return JSON.stringify({ httpStatus: 200, message: `Appointment with ID ${appointmentId} updated successfully.`, appointment: updatedAppointment });
     } catch (error) {
@@ -282,7 +271,6 @@ mqttReq.response("v1/appointments/update", (payload) => {
         return JSON.stringify({ httpStatus: 500, message: 'Internal Server Error' });
     }
 });
-
 
 client.on("connect", () => {
     console.log("scheduling-service connected to broker")
